requirements.txt
langchain
langchain-community
streamlit
pypdf
PyPDF2
pdf2image
numpy
python-dotenv
tqdm
faiss-cpu
langchain-ollama
tree